Bank System
This is a program that allows its users to create bank accounts and do basic banking operations. Note that this program had to be restricted to the use of the following java topics: Elementary Programming - Selections - Mathematical Functions - Strings - Characters - Loops - Methods - Arrays. It was completed on the 12th of February 2025.