Add Feed The Monster Game Program
GUI-based program that uses JavaFX. The game begins by showing a window with a monster-looking node in the middle. The monster has to be clicked five times. Every time it is clicked, it is fed, so its size grows. In addition, with every click on the monster, its color changes randomly and its location on the window changes randomly too. After the five clicks, a text node appears. The message displays the time elapsed from the first click to the last. On the sixth click, the game resets. 
Note that the game/program is run from the FeedTheMonsterGame class.